Chords and layers #70
Enhance button handling with a hierarchical state machine.
The external state machine processes the button state to select which mapping to activate and the internal state machine raises events and call actions registered in the new mapping structure.

This change only attempts to replicate the existing behaviour in the new logic that will open the possibility for advanced button mapping such as toggles, turbo, chords layers, etc... Currently I need to rework the Mapping structure to take less memory because I bust the Pico's limit right now.

Sticky is the only button mode I don't have an implementation for yet, so I am keeping some legacy code for it.